dataset/requirements.txt
soundfile
sounddevice
numpy
sox
python-cfonts
nltk
keyboard
soundcard
websockets
